refactor(capture/imap): remove the dead transient IMAP flavor
imap.New() (the standing:false one-shot puller Connector.Authenticate/
Sync/HealthCheck branched on) and its Stats() accessor had no
production caller left: ConnectImap/RunTransient were deleted earlier
in this branch, and only imap.NewStanding() is wired into the capture
registry (internal/compose/capture.go, connectors_imap.go). Only tests
exercised the transient path.

Removes New(), Close(), the conn/netConn fields it managed,
syncTransient, the authConfig type it alone used, the now-unused
readMessageBody wrapper, the standing bool field, and the three
if c.standing branches in Authenticate/Sync/HealthCheck -- each method
now calls straight through to its standing-only implementation. Keeps
the Stats type (still written internally by capture/captureFetched for
per-pull bookkeeping) but drops its Connector-level accessor and the
Mailbox field that only the deleted transient path ever set, and
re-documents it as internal-only.

Deletes transient_sync_test.go (both tests were transient-path-only,
now orphaned). Salvages TestDescriptorDeclaresReadOnlyCapture and
TestBoundedWindowClamps into imap_test.go against NewStanding(), since
both exercise still-live production code. Drops
TestCloseIsSafeWithoutSession (tested the removed Close()).

Corrects the four comments that described a transient flavor no longer
in the tree: standing.go's NewStanding doc ("the transient one-shot
pull keeps using New()"), this package's top-of-file doc and the
Connector struct doc (both described the one-shot/standing-bool shape),
and frontend/src/screens/imap-connect-form.tsx's doc, which called the
onboarding IMAP panel "a separate, still-unmigrated surface" -- it
posts to the exact same standing /connectors/imap/connect endpoint
(verified against onboarding-connect-panels.tsx's ImapConnectPanel).

Verified via go build -tags integration ./... and the full imap package
test suite; a repo-wide grep for imap.New( and imap .Stats() came back
empty outside this diff.
